requirements.txt
fastmcp>=0.1.0
anthropic>=0.39.0
openai>=1.12.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
requests>=2.31.0
python-telegram-bot>=20.0
fastapi>=0.104.0
uvicorn>=0.24.0
